feat: Implement rendering systems for board, entities, snakes, and UI
- Added BoardRenderSystem to handle rendering of the game board, including background, grid, and basic tiles.
- Introduced EntityRenderSystem for rendering generic entities based on Position and Renderable components.
- Created SnakeRenderSystem for rendering snake entities with smooth interpolation effects.
- Developed UIRenderSystem to manage UI overlays such as score, speed bar, and music indicators.
- Updated GameplayScene to integrate new rendering systems and ensure proper rendering order.
- Added ColorScheme entity to provide a global color palette for rendering systems.
